+++ user_profile_update.php
@@ -131,8 +131,8 @@ if (!empty($_FILES['image']['name'])) {
         if ($update_image) {
             move_uploaded_file($image_tmp_name, $image_folder . $image);
 
-            // Delete the old image if it's not the default image
-            if ($old_image !== 'default.png') {
+            // Delete the old image if it exists and it's not the default image
+            if ($old_image !== 'default.png' && file_exists($image_folder . $old_image)) {
                 unlink($image_folder . $old_image);
             }
